+++ docs/src/components/spark-ui/glare-hover/glare-hover.vue
@@ -61,9 +61,9 @@ const rootClass = computed(() =>
     "before:[background-position:-100%_-100%,0_0]",
     // TRANSITION
     !props.playOnce &&
-      "before:transition-[background-position] before:duration-[var(--gh-duration)] before:ease-in-out",
+      "before:[transition-property:background-position] before:[transition-duration:var(--gh-duration)] before:ease-in-out",
     props.playOnce &&
-      "before:transition-none hover:before:transition-[background-position] hover:before:duration-[var(--gh-duration)]",
+      "before:[transition-property:none] hover:before:[transition-property:background-position] hover:before:[transition-duration:var(--gh-duration)] hover:before:ease-in-out",
     // HOVER EFFECT
     "hover:before:[background-position:100%_100%,0_0]",
     props.class,